fix(analytics): apply the framing fallback to the analytics scripts too
Only the consent banner had the client-side check. SpeedInsights and
GoogleAnalytics were gated purely on Sec-Fetch-Dest, so on exactly the
browsers and proxies the fallback exists for, the banner hid itself but
the scripts still mounted in the frame and reported the visit twice.

Hoist the check into TopLevelOnly and wrap all three. Both script
components are client components that inject on mount rather than into
the server HTML, so a framed document never runs them even when the
header is missing and the markup was sent.

Also note in embed.ts that this is not the same signal as the embed=true
search param demo-content.tsx reads, since only a page can see search
params and the two answer different questions.
