+++ src/lib/embed.ts
@@ -11,7 +11,12 @@ import { headers } from "next/headers";
  * sets to `iframe` on the framed navigation and `document` on the top-level
  * one. A request that arrives without the header (Safari below 16.4, or a
  * proxy that strips it) reads as not-embedded, so the caller falls back to
- * rendering exactly what it renders today.
+ * rendering the markup and `<TopLevelOnly>` drops it on the client instead.
+ *
+ * Not the same signal as the `embed=true` search param that `demo-content.tsx`
+ * reads. That one is the demo asking for its own chrome to be hidden, and only
+ * a page can see search params; this one is the layout asking whether it is the
+ * document the user is actually looking at. They are deliberately separate.
  */
 export async function isEmbeddedDocument(): Promise<boolean> {
   const requestHeaders = await headers();
